CHG: Make input checks optional
We use a template parameter to make input checks optional.
User facing functions have the checks enabled. Internal function
calls when bootstrapping don't have the checks enabled as we can ensure
correctness.

This means the ifault parameter has been changed to proper runtime
extensions.
